+++ common/mods/necron/stargate/BlockCoord.java
@@ -16,7 +16,7 @@ public abstract class BlockCoord extends BlockGui {
 	 */
 	@Override
 	public void onBlockPlacedBy(World world, int x, int y, int z, EntityLiving entityLiving) {
-		int angle = MathHelper.floor_double((double) (entityLiving.rotationYaw * 4.0F / 360.0F) + 0.5D) & 3;
+		int angle = MathHelper.floor_double(entityLiving.rotationYaw * 4.0F / 360.0F + 0.5D) & 3;
 		
 		switch(angle) {
 			case 0:
